hook作成
