+++ includes/render.h
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/03 03:18:31 by tsishika          #+#    #+#             */
-/*   Updated: 2024/04/03 03:49:55 by tsishika         ###   ########.fr       */
+/*   Updated: 2024/04/03 04:03:31 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,11 @@
 # include "mlx.h"
 # include "struct.h"
 
-// my_mlx_init.c
+// hook.c
+int		key_press(int keycode, t_data *data);
+int		close_window(t_data *data);
+
+// init.c
 void	my_mlx_init(t_data *data);
 
 // my_mlx_pixel_put.c
